pyproject.toml
[project]
name = "vnpt-track2"
version = "2.0.0"
description = "Titan Shield RAG System - Vietnamese QA for VNPT Hackathon Track 2"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=2.9.0",
    "aiohttp>=3.8.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "bm25s>=0.1.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
developer = []
development = [
    "datasets>=4.4.1",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.5.0",
    "loguru>=0.7.3",
    "mlflow>=3.7.0",
    "openai>=2.9.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ratelimit>=2.2.1",
]
